Moves "isNodeProcess" into internal utilities
