+++ src/setupWorker/setupWorker.ts
@@ -2,7 +2,7 @@ import { SetupWorkerInternalContext, RequestHandlersList } from './glossary'
 import { createStart } from './start/createStart'
 import { createStop } from './stop/createStop'
 import * as requestHandlerUtils from '../utils/requestHandlerUtils'
-import { isNodeProcess } from '../utils/isNodeProcess'
+import { isNodeProcess } from '../utils/internal/isNodeProcess'
 import { ServiceWorkerMessage } from '../utils/createBroadcastChannel'
 
 export interface SetupWorkerApi {
