+++ src/utils/internal/isNodeProcess.node.test.ts
